the whole system of Skills is gone, skilledPlayers are here instead.
Players and enemies both got the Map property set during the Map constructor.

Movement should now work, but it was not tested.

Everything about input is now solved in the Game class ( esc no longer in the InputGetter).

operator + in Location added

Useless method ShowThis in GameControl deleted.

InputGetter.CurrButton is now queue instead of a single button.

Opening In-game-menu (and closing it) now stops/starts time and it is solved in the InGameMenu class.

All methods from Map class that determines wether an object may go ina direction were moved to inner class DirectionAccesor.

Map Cloning is back here and it depends on serialization and deserialization.

MapBuilders now gets a stream in a constructor, not as a parameter of GetMap method.

MapSaver now exists... but it does nothing yet.

Specification file added and send to my teacher.